requirements.txt
PySide6>=6.5
pillow>=10.0
pandas>=2.0
numpy>=1.23
scikit-image>=0.21
